requirements.txt
# Shrimp Pond RAG Pipeline - Dependencies

# Core RAG Framework
langchain==0.1.0
langchain-community==0.0.13
langchain-core==0.1.0
langchain-text-splitters==0.0.1

# Vector Store and Embeddings
chromadb==0.4.22
sentence-transformers==2.3.1

# LLM Integration
openai==1.10.0

# Data Processing
numpy==1.24.3
pandas==2.0.3

# Configuration and Logging
pyyaml==6.0.1
python-dotenv==1.0.0

# Web UI Framework
streamlit==1.31.0

# Utilities
tqdm==4.66.1
requests==2.31.0
